# src/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: REST API DOC
  description: |
    REST API 문서
  version: 1.0.0
servers:
  - url: 'http://192.168.0.86:4000/api'
paths:
...
  /users/exist:
    post:
      tags:
        - User
      summary: 유저가 존재하는지 확인(로그인 시 무조건 실행해야 함)
      description: 유저가 존재하면 isExist 가 true, 존재하지 않으면 false</br>
                  isExist->true 면 access_token 을 뱉어주는데 이 access_token 은 BearerAuth 에서 사용될 토큰임</br>
                  isExist->false 면 access_token 을 주지 않고 users/add 라우트를 실행해야 함
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                login_type:
                  type: integer
                  enum:
                    - 1
                    - 2
                  description: 1 -> naver</br>2 -> kakao
                access_token:
                  type: string
                  description: 로그인 버튼 누르고 나오는 url 에서 받아오는 access_token 임
      responses:
        '200':
          description: response 가 200 && 이미 유저가 존재할 떄 </br> 또한 access_token 은 bearerAuth 에 넣을 토큰임
          content:
            application/json:
              schema:
                type: object
                example:
                  status: 200
                  data:
                    Auth: 1
                    isExist: true
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        '400':
          $ref: '#/components/responses/400'
  /users:
    post:
      tags:
        - User
      summary: exist api 요청 이후 수행하는 등록 api
      description: /users/exist 에서 isExist 가 false 면 실행되는 api 이다.</br>
                  여기서 나머지 필드를 다 채우고 정확한 값을 요청하면 status 200 과</br>
                  BearerAuth 에서 사용된 accescc_token 을 함께 띄워준다.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                login_type:
                  type: integer
                  enum:
                    - 1
                    - 2
                  default: 2
                  description: 1 -> naver</br>2 -> kakao
                access_token:
                  type: string
                  description: /users/isExist 의 body 에서 받아온 token
                phone:
                  type: string
                  description: 회원의 전화번호
                name:
                  type: string
                  description: 회원의 이름
                email:
                  type: string
                  description: 회원의 이메일
                student_ID:
                  type: string
                  description: 회원의 학번
              required:
                - login_type
                - access_token
                - phone
                - name
                - email
                - student_ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                example:
                  status: 200
                  data:
                    userToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    '400':
      description: bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: integer
              msg:
                type: string
            example:
              status: 400
              msg: bad request
    '401':
      description: invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: integer
              msg:
                type: string
            example:
              status: 401
              msg: 인증 오류
...